requirements.txt
asgiref==3.8.1
boto3==1.34.140
botocore==1.34.162
certifi==2025.6.15
charset-normalizer==3.4.2
dj-database-url==3.0.0
Django==5.2.3
django-cors-headers==4.7.0
django-sendgrid-v5==1.3.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docopt==0.6.2
ecdsa==0.19.1
git-filter-repo==2.47.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
MarkupSafe==3.0.2
mercadopago==2.3.0
packaging==25.0
pillow==11.2.1
pipreqs==0.4.13
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
requests==2.32.4
s3transfer==0.10.4
sendgrid==6.12.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
user-agents==2.2.0
Werkzeug==3.1.3
whitenoise==6.9.0
yarg==0.1.10
